fix: remove background map
