+++ pages/index.vue
@@ -7,6 +7,8 @@ definePageMeta({
 <template>
 	<MainContent>
 		<h1 class="sr-only">VICAV - Vienna Corpus of Arabic Varieties</h1>
-		<AppBackground />
+		<div class="relative isolate grid h-full w-full">
+			<WindowManager />
+		</div>
 	</MainContent>
 </template>
